refactor: extract route change indicator to its own component
